fix issue where BLEPeripheral lib needed to be included even if not needed
